dashboard/app.py
import streamlit as st
import pandas as pd
import plotly.graph_objects as go
import snowflake.connector
from datetime import datetime
from functools import lru_cache
from dotenv import load_dotenv
import os