requirements.txt
kafka-python==2.0.2
boto3==1.34.14

# Snowflake connector + DBT compatible versions
snowflake-connector-python==3.10.1
dbt-core==1.7.4
dbt-snowflake==1.7.1

# Utils
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.1
streamlit
pandas
plotly
streamlit-aggrid  # optional for table viewer
